+++ demo/DemoIsolated/Model/Base/IsoNullableEnum.generated.cs
@@ -1,0 +1,11 @@
+using System.Runtime.Serialization;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace DemoIsolated.Model.Base {
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum IsoNullableEnum {
+        [EnumMember(Value = "first")] First,
+        [EnumMember(Value = "second")] Second
+    }
+}