+++ demo/DemoIsolated/Model/CustomTimeConverter.cs
@@ -1,0 +1,8 @@
+using Caffoa.JsonConverter;
+
+namespace DemoIsolated.Model;
+
+public class CustomTimeConverter : CaffoaTimeOnlyConverter
+{
+    
+}