+++ demo/DemoIsolated/Model/IsoError.generated.cs
@@ -1,0 +1,39 @@
+#pragma warning disable CS0612
+#pragma warning disable CS0618
+
+using System;
+using Caffoa;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
+using System.Linq;
+using DemoIsolated.Model.Base;
+
+namespace DemoIsolated.Model {
+/// AUTOGENERED BY caffoa ///
+    public partial class IsoError {
+        public const string IsoErrorObjectName = "error";
+        /// <summary>
+        /// Single string based code describing the error.
+        /// </summary>
+        [JsonProperty("status", Required = Required.Always)]
+        public virtual string Status { get; set; }
+
+        /// <summary>
+        /// Human readable error message.
+        /// </summary>
+        [JsonProperty("message", Required = Required.Always)]
+        public virtual string Message { get; set; }
+
+        [JsonExtensionData]
+        public Dictionary<string, object> AdditionalProperties;
+
+        public IsoError(){}
+        public IsoError(IsoError other) {
+            Status = other.Status;
+            Message = other.Message;
+            AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, object>(other.AdditionalProperties) : null;
+        }
+        public IsoError ToIsoError() => new IsoError(this);
+    }
+}