+++ demo/DemoIsolated/Model/IsoExtendedAddress.AddressTypeValue.generated.cs
@@ -1,0 +1,13 @@
+using System.Runtime.Serialization;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace DemoIsolated.Model {
+    public partial class IsoExtendedAddress {
+        // enum values for "addressType"
+        [JsonConverter(typeof(StringEnumConverter))]
+        public enum AddressTypeValue {
+            [EnumMember(Value = "regular")] Regular
+        }
+    }
+}