+++ demo/DemoIsolated/Model/IsoPricing.generated.cs
@@ -1,0 +1,34 @@
+#pragma warning disable CS0612
+#pragma warning disable CS0618
+
+using System;
+using Caffoa;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
+using System.Linq;
+using System.Collections.Immutable;
+using DemoIsolated.Model.Base;
+
+namespace DemoIsolated.Model {
+/// AUTOGENERED BY caffoa ///
+    public partial class IsoPricing {
+        public const string IsoPricingObjectName = "pricing";
+        [JsonProperty("price")]
+        public virtual double? Price { get; set; }
+
+        [JsonProperty("taxes")]
+        public virtual Dictionary<string, double> Taxes { get; set; } = new Dictionary<string, double>();
+
+        [JsonExtensionData]
+        public Dictionary<string, object> AdditionalProperties;
+
+        public IsoPricing(){}
+        public IsoPricing(IsoPricing other) {
+            Price = other.Price;
+            Taxes = other.Taxes?.ToDictionary(entry => entry.Key, entry => entry.Value);
+            AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, object>(other.AdditionalProperties) : null;
+        }
+        public IsoPricing ToIsoPricing() => new IsoPricing(this);
+    }
+}