+++ demo/DemoIsolated/Model/IsoTagInfos.generated.cs
@@ -1,0 +1,29 @@
+#pragma warning disable CS0612
+#pragma warning disable CS0618
+
+using System;
+using Caffoa;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
+using System.Linq;
+using DemoIsolated.Model.Base;
+
+namespace DemoIsolated.Model {
+/// AUTOGENERED BY caffoa ///
+    public partial class IsoTagInfos {
+        public const string IsoTagInfosObjectName = "tagInfos";
+        [JsonProperty("user")]
+        public virtual Dictionary<string, List<Guid>> User { get; set; } = new Dictionary<string, List<Guid>>();
+
+        [JsonExtensionData]
+        public Dictionary<string, object> AdditionalProperties;
+
+        public IsoTagInfos(){}
+        public IsoTagInfos(IsoTagInfos other) {
+            User = other.User?.ToDictionary(entry => entry.Key, entry => entry.Value);
+            AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, object>(other.AdditionalProperties) : null;
+        }
+        public IsoTagInfos ToIsoTagInfos() => new IsoTagInfos(this);
+    }
+}