+++ demo/DemoIsolated/Model/IsoUser.generated.cs
@@ -1,0 +1,77 @@
+#pragma warning disable CS0612
+#pragma warning disable CS0618
+
+using System;
+using Caffoa;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System.Collections.Generic;
+using System.Linq;
+using System.Collections.Immutable;
+using Caffoa.JsonConverter;
+using DemoIsolated.Model.Base;
+
+namespace DemoIsolated.Model {
+/// AUTOGENERED BY caffoa ///
+    public partial class IsoUser : IsoAnyUser {
+        public const string IsoUserObjectName = "user";
+        /// <summary>
+        /// A fancy string with description
+        /// </summary>
+        [JsonProperty("name", Required = Required.Always)]
+        public virtual string Name { get; set; }
+
+        [JsonProperty("address")]
+        public virtual IsoAddress Address { get; set; }
+
+        [Obsolete]
+        [JsonConverter(typeof(CaffoaDateOnlyConverter))]
+        [JsonProperty("birthdate")]
+        public virtual DateOnly? Birthdate { get; set; }
+
+        [JsonProperty("emails")]
+        public virtual ICollection<string> Emails { get; set; } = new List<string>();
+
+        [JsonProperty("descriptions")]
+        public virtual Dictionary<string, string> Descriptions { get; set; } = new Dictionary<string, string>();
+
+        [JsonProperty("type", Required = Required.Always)]
+        public virtual TypeValue Type { get; set; } = TypeValue.Simple;
+
+        [JsonProperty("role")]
+        public virtual RoleValue Role { get; set; } = RoleValue.Reader;
+
+        [Obsolete("do not use this")]
+        [JsonProperty("ageGroup")]
+        public virtual int? AgeGroup { get; set; } = 40;
+
+        [Obsolete("do not use this")]
+        [JsonConverter(typeof(CustomTimeConverter))]
+        [JsonProperty("preferredContactTime")]
+        public virtual TimeOnly PreferredContactTime { get; set; } = TimeOnly.Parse("12:00");
+
+        [JsonProperty("lastSessionLength")]
+        public virtual TimeSpan LastSessionLength { get; set; }
+
+        [JsonExtensionData]
+        public Dictionary<string, object> AdditionalProperties;
+
+        public IsoUser(){}
+        public IsoUser(IsoUser other) {
+            Name = other.Name;
+            Address = other.Address?.ToIsoAddress();
+            Birthdate = other.Birthdate;
+            Emails = other.Emails?.ToList();
+            Descriptions = other.Descriptions?.ToDictionary(entry => entry.Key, entry => entry.Value);
+            Type = (IsoUser.TypeValue)other.Type;
+            Role = (IsoUser.RoleValue)other.Role;
+            AgeGroup = other.AgeGroup;
+            PreferredContactTime = other.PreferredContactTime;
+            LastSessionLength = other.LastSessionLength;
+            AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, object>(other.AdditionalProperties) : null;
+        }
+        public IsoUser ToIsoUser() => new IsoUser(this);
+        public virtual IsoAnyUser ToIsoAnyUser() => ToIsoUser();
+        public virtual string TypeDiscriminator => Type.Value();
+    }
+}