+++ demo/DemoV1b/DemoV1bMaintainanceFunctions.generated.cs
@@ -7,10 +7,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.Extensions.Logging;
-
 using Microsoft.Azure.WebJobs.Extensions.DurableTask;
 using DemoV1b.Model.Base;
 using DemoV1b.Model;